excludes category and prohibits the exclusion of a category linked to another place

+++ src/pages/management.php
@@ -12,7 +12,13 @@ if (isset($_GET['admin'])) :
       $produtos->deleteProdutoCategoria($_GET['id']);
       $produtos->delete($_GET['id']);
     } elseif ($_GET['delete'] == "categoria") {
-      $categorias->delete($_GET['id']);
+      $response = $categorias->delete($_GET['id']);
+      if ($response != true) {
+        echo '<script>
+          alert("Aconteceu algo de errado! Você deve estar tentando excluír uma categoria vinculada a outros registros!");
+          window.location="management.php?admin=categorias";
+        </script>';
+      }
     }
   }
 ?>
@@ -142,7 +148,7 @@ if (isset($_GET['admin'])) :
                       <path fill-rule="evenodd" d="M1 13.5A1.5 1.5 0 0 0 2.5 15h11a1.5 1.5 0 0 0 1.5-1.5v-6a.5.5 0 0 0-1 0v6a.5.5 0 0 1-.5.5h-11a.5.5 0 0 1-.5-.5v-11a.5.5 0 0 1 .5-.5H9a.5.5 0 0 0 0-1H2.5A1.5 1.5 0 0 0 1 2.5v11z" />
                     </svg>
                   </a>
-                  <a href="?admin=categorias&delete=categoria&id=<?php echo $value['cd'] ?>" title="Deletar">
+                  <a href="?admin=categorias&delete=categoria&id=<?php echo $value['cd'] ?>" onclick="return confirm('Tem certeza que deseja deletar este registro?')" title="Deletar">
                     <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-trash" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                       <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5zm3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0V6z" />
                       <path fill-rule="evenodd" d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1v1zM4.118 4L4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z" />
